main/requirements.txt
# Function dependencies, for example:
# package>=version
earthengine-api
google-cloud-storage